tests/test_image_dedup.py
"""Tests for perceptual image deduplication."""

import pytest
from pathlib import Path

from src.analysis.image_dedup import (
    is_image,
    hamming_distance,
    IMAGE_EXTENSIONS,
)


class TestIsImage:
    def test_jpg(self):
        assert is_image("/test/photo.jpg") is True

    def test_jpeg(self):
        assert is_image("/test/photo.jpeg") is True

    def test_png(self):
        assert is_image("/test/icon.png") is True

    def test_gif(self):
        assert is_image("/test/anim.gif") is True

    def test_webp(self):
        assert is_image("/test/img.webp") is True

    def test_bmp(self):
        assert is_image("/test/img.bmp") is True

    def test_tiff(self):
        assert is_image("/test/img.tiff") is True

    def test_non_image_txt(self):
        assert is_image("/test/doc.txt") is False

    def test_non_image_pdf(self):
        assert is_image("/test/doc.pdf") is False

    def test_non_image_py(self):
        assert is_image("/test/code.py") is False

    def test_case_insensitive(self):
        assert is_image("/test/photo.JPG") is True
        assert is_image("/test/photo.Png") is True


class TestHammingDistance:
    def test_identical_hashes(self):
        """Identical hashes have distance 0."""
        assert hamming_distance("abcd1234", "abcd1234") == 0

    def test_completely_different(self):
        """Completely different hashes have high distance."""
        d = hamming_distance("0000", "ffff")
        assert d > 0

    def test_one_bit_difference(self):
        """Single bit difference gives small distance."""
        # 0x0 = 0000, 0x1 = 0001 -> 1 bit different
        d = hamming_distance("0", "1")
        assert d == 1

    def test_symmetric(self):
        """Hamming distance is symmetric."""
        d1 = hamming_distance("abcd", "1234")
        d2 = hamming_distance("1234", "abcd")
        assert d1 == d2


class TestImageExtensions:
    def test_common_extensions_included(self):
        """Common image formats are in the set."""
        expected = {".jpg", ".jpeg", ".png", ".gif", ".bmp", ".webp"}
        assert expected.issubset(IMAGE_EXTENSIONS)


class TestFindSimilarImages:
    def test_empty_list(self):
        """Empty input returns empty output."""
        from src.analysis.image_dedup import find_similar_images

        result = find_similar_images([])
        assert result == []

    def test_nonexistent_files(self):
        """Nonexistent files are skipped gracefully."""
        from src.analysis.image_dedup import find_similar_images

        result = find_similar_images(["/nonexistent/a.jpg", "/nonexistent/b.jpg"])
        assert result == []

    def test_with_real_identical_images(self, tmp_path):
        """Identical images should be grouped together."""
        try:
            from PIL import Image
        except ImportError:
            pytest.skip("Pillow not installed")

        # Create two identical small images
        img = Image.new("RGB", (100, 100), color="red")
        path1 = str(tmp_path / "img1.png")
        path2 = str(tmp_path / "img2.png")
        img.save(path1)
        img.save(path2)

        from src.analysis.image_dedup import find_similar_images

        result = find_similar_images([path1, path2], threshold=10)
        assert len(result) == 1
        assert result[0]["count"] == 2

    def test_with_different_images(self, tmp_path):
        """Very different images should not be grouped."""
        try:
            from PIL import Image
        except ImportError:
            pytest.skip("Pillow not installed")

        # Create two very different images
        img1 = Image.new("RGB", (100, 100), color="red")
        img2 = Image.new("RGB", (100, 100), color="blue")

        path1 = str(tmp_path / "red.png")
        path2 = str(tmp_path / "blue.png")
        img1.save(path1)
        img2.save(path2)

        from src.analysis.image_dedup import find_similar_images

        # With strict threshold, solid colors may still match depending on the hash
        # Use threshold=0 for exact match only
        result = find_similar_images([path1, path2], threshold=0)
        # They should either be separate or have distance > 0
        if result:
            assert all(g["count"] >= 2 for g in result)
