pyproject.toml
[project]
name = "doc-intelligence"
version = "5.0.0"
description = "AI-powered file intelligence — persistent indexing, smart tagging, natural language queries, health reports, and Streamlit dashboard"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "duckdb>=0.10.0",
    "xxhash>=3.4.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "pypdf>=3.0.0",
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
images = [
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
]
dashboard = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
]
ai = [
    "anthropic>=0.39.0",
]
openai = [
    "openai>=1.0.0",
]
voyage = [
    "voyageai>=0.3.0",
]
docs = [
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
]
mcp = [
    "mcp>=1.0.0",
]
all = [
    "doc-intelligence[images,dashboard,ai,openai,voyage,docs,mcp]",
]

[project.scripts]
doc-intelligence = "scripts.cli:main"
doc-scan = "scripts.scan:main"
doc-search = "scripts.search:main"
doc-duplicates = "scripts.find_duplicates:main"
doc-cleanup = "scripts.cleanup:main"
doc-watch = "scripts.watch:main"
doc-big-files = "scripts.big_files:main"
doc-report = "scripts.report:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
